requirements.txt
langchain==0.0.250
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.25.0
faiss-cpu==1.7.4
streamlit-extras
google-auth==2.22.0
google-cloud-aiplatform==1.30.1
streamlit-chat==0.1.1
tiktoken==0.4.0
transformers==4.34.0
